offline support

+++ scripts/dbhandler.js
@@ -1,8 +1,10 @@
 import sqlite from 'better-sqlite3';
-import {writeFile, mkdir} from 'fs';
+import {writeFile, mkdir, existsSync} from 'fs';
 
-mkdir('./data', () => {});
-writeFile('./data/main.db', '', () => {});
+if (!existsSync('./data/main.db')) {
+    mkdir('./data', () => {});
+    writeFile('./data/main.db', '', () => {});
+}
 const db = sqlite('./data/main.db');
 
 db.prepare('CREATE TABLE IF NOT EXISTS accounts (id INTEGER PRIMARY KEY AUTOINCREMENT, username STRING, password STRING, topScore INT)').run();
